Renamed all classes to CM... to avoid Flutter 3 conflicts.
This package has not been updated by the publisher in 2 years. The recent release of flutter 3 'stole' the name MenuItem, therefore all code using this package has conflicts. I branched the code and made it our own CrowdMarket version by prepending CM on the front of all classes.  We are now responsible for this package going forward.